requirements.txt
Flask==2.2.2
requests==2.28.1
python-dotenv==0.20.0
Werkzeug<3.0